virtualisation module connected

+++ app/graphql/resolvers/VM/delete.js
@@ -2,6 +2,7 @@ import {AuthForBoth} from "../../middleWare.js"
 import {PrismaClient} from "@prisma/client"
 const prisma = new PrismaClient();
 import { GraphQLError } from "graphql";
+import { createCall } from "../Virtualization/index.js";
 
 //// For Delete Virtual Machine //////////////
 
@@ -44,22 +45,33 @@ const deleteVMResolvers = {
               console.log({ in: id });
               console.log(i);
               if (forDeleteVM) {
-                const forNot = await prisma.notification.deleteMany({
-                  where: {
-                    vmId: { in: id },
-                  },
-                });
-                console.log(forNot);
-                const deleteVMId = await prisma.virtualMachine.deleteMany({
-                  where: {
-                    id: { in: id },
-                  },
-                });
-                console.log(deleteVMId);
-                if (deleteVMId) {
-                  console.log("del");
+                let result = await createCall("deleteCall",{ "name": forDeleteVM[i].virtualMachineName})
+                if(result.data["result"]["status"]){
+                  const forNot = await prisma.notification.deleteMany({
+                    where: {
+                      vmId: { in: id },
+                    },
+                  });
+                  console.log(forNot);
+                  const deleteVMId = await prisma.virtualMachine.deleteMany({
+                    where: {
+                      id: { in: id },
+                    },
+                  });
+                  console.log(deleteVMId);
+                  if (deleteVMId) {
+                    console.log("del");
+                  }
+                  return "VM_Deleted";
                 }
-                return "VM_Deleted";
+                else{
+                  throw new GraphQLError('Failed to Delete', {
+                    extensions: {
+                      StatusCode: 404,
+                      code: 'Failed '
+                    }
+                  }) 
+                }
               } else {
                 throw new Error("error");
               }
@@ -92,14 +104,14 @@ const deleteVMResolvers = {
                 },
               },
             });
-// console.log(forDeleteVM.User , "kkkk");
-// console.log(forDeleteVM.userId , lll);
-console.log(forDeleteVM);
+            console.log(forDeleteVM);
             console.log(forDeleteVM.length);
             for (let j = forDeleteVM.length; j >= 0; j++) {
               console.log({ in: id });
               console.log(j);
               if (forDeleteVM) {
+                let result = await createCall("deleteCall",{ "name": forDeleteVM[i].virtualMachineName})
+                if(result.data["result"]["status"]){
                 const forNot = await prisma.notification.deleteMany({
                   where: {
                     vmId: { in: id },
@@ -115,6 +127,15 @@ console.log(forDeleteVM);
                 });
                 console.log(deleteVMId.userId);
                 return "VM_Deleted";
+              }
+              else{
+                throw new GraphQLError('Failed to Delete', {
+                  extensions: {
+                    StatusCode: 404,
+                    code: 'Failed '
+                  }
+                }) 
+              }
               } else {
                 throw new Error("error");
               }
